Add `Sai::Conversion::RGB::ColorSpace`
